+++ app/src/main/java/com/google/codelab/hotpepperapiapp/viewModel/StoreListViewModel.kt
@@ -1,9 +1,8 @@
 package com.google.codelab.hotpepperapiapp.viewModel
 
 import androidx.lifecycle.ViewModel
-import com.google.codelab.hotpepperapiapp.Failure
-import com.google.codelab.hotpepperapiapp.FailureType
-import com.google.codelab.hotpepperapiapp.getMessage
+import com.google.codelab.hotpepperapiapp.model.Failure
+import com.google.codelab.hotpepperapiapp.model.getMessage
 import com.google.codelab.hotpepperapiapp.model.response.StoresResponse
 import com.google.codelab.hotpepperapiapp.usecase.StoreListUseCase
 import com.google.codelab.hotpepperapiapp.usecase.StoreListUseCaseImpl
